+++ src/components/molecules/FormEditor/FormEditor.tsx
@@ -14,10 +14,10 @@ import {stringify} from 'yaml';
 
 import {DEFAULT_EDITOR_DEBOUNCE} from '@constants/constants';
 
-import {settingsSelector} from '@redux/appConfig';
+import {isInClusterModeSelector, settingsSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAutosavingStatus} from '@redux/reducers/main';
-import {isInClusterModeSelector, isInPreviewModeSelectorNew, selectedFilePathSelector} from '@redux/selectors';
+import {isInPreviewModeSelectorNew, selectedFilePathSelector} from '@redux/selectors';
 import {useSelectedResource} from '@redux/selectors/resourceSelectors';
 import {mergeManifests} from '@redux/services/manifest-utils';
 import {removeSchemaDefaults} from '@redux/services/schema';
