+++ src/components/organisms/ActionsPane/Scale/Scale.tsx
@@ -5,10 +5,9 @@ import {Col, InputNumber, Modal, Row, Tooltip} from 'antd';
 import {TOOLTIP_DELAY} from '@constants/constants';
 import {ScaleTooltip} from '@constants/tooltips';
 
-import {kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/appConfig';
+import {isInClusterModeSelector, kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {closeScaleModal, openScaleModal} from '@redux/reducers/ui';
-import {isInClusterModeSelector} from '@redux/selectors';
 import {useSelectedResource} from '@redux/selectors/resourceSelectors';
 import scaleDeployment from '@redux/services/scaleDeployment';
 import {startClusterConnection} from '@redux/thunks/cluster';
