+++ src/components/organisms_new/PaneManager/PaneManager.tsx
@@ -2,10 +2,9 @@ import React, {useCallback, useMemo} from 'react';
 
 import {Skeleton} from 'antd';
 
-import {activeProjectSelector} from '@redux/appConfig';
+import {activeProjectSelector, isInClusterModeSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setPaneConfiguration, toggleLeftMenu} from '@redux/reducers/ui';
-import {isInClusterModeSelector} from '@redux/selectors';
 
 import {ActionsPane, BottomPaneManager, Dashboard, GitOpsView, NavigatorPane} from '@organisms';
 import {EmptyDashboard} from '@organisms/Dashboard/EmptyDashboard';
