+++ src/navsections/ImagesSectionBlueprint/ImagesQuickAction.tsx
@@ -1,6 +1,7 @@
+import {isInClusterModeSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {openReplaceImageModal} from '@redux/reducers/ui';
-import {isInClusterModeSelector, isInPreviewModeSelectorNew} from '@redux/selectors';
+import {isInPreviewModeSelectorNew} from '@redux/selectors';
 
 import {ItemCustomComponentProps} from '@shared/models/navigator';
